requirements.txt
alembic==1.14.1
annotated-types==0.7.0
apted==1.0.3
black==25.1.0
boto3==1.37.3
botocore==1.37.3
bottle==0.13.2
Brotli==1.1.0
CachedMethods==0.1.4
certifi==2025.1.31
CGRtools==4.1.35
charset-normalizer==3.4.1
cli_exit_tools==1.2.7
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
dask==2025.2.0
Deprecated==1.2.18
dill==0.3.9
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
greenlet==3.1.1
huggingface-hub==0.17.3
idna==3.10
importlib_metadata==8.6.1
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
lib-detect-testenv==2.0.8
locket==1.0.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
metaflow==2.15.0
multiprocess==0.70.17
multivolumefile==0.2.3
mypy-extensions==1.0.0
numpy==1.26.4
optuna==3.6.1
optuna-dashboard==0.16.0
packaging==24.2
pandas==2.2.3
partd==1.4.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
psutil==7.0.0
py7zr==0.18.12
pyarrow==19.0.1
pybcj==1.0.3
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyppmd==0.18.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzstd==0.16.2
rdchiral==1.1.0
rdkit==2023.9.6
reaction_utils==1.8.0
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.38
swifter==1.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tokenizers==0.14.1
toolz==1.0.0
torch==1.13.1+cu117
tqdm==4.67.1
transformers==4.35.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wrapt==1.17.2
wrapt_timeout_decorator==1.5.1
xxhash==2.0.2
zipfile-deflate64==0.2.0
zipp==3.21.0
